Sorted out tests.
